+++ app/index.js
@@ -1,3 +1,5 @@
+import Avatar from './avatar.js';
+
 var USER_DATA = {
 	name: "Dylan SHWOOPS",
 	username: "dylanshwerps",
@@ -7,56 +9,48 @@ var USER_DATA = {
 var React = require('react');
 var ReactDOM = require('react-dom');
 
-// var HelloWorld = React.createClass({
-// 	render: function() {
+// var ProfilePic = React.createClass({
+// 	render: function () {
+// 		return <img src={this.props.imageUrl} style={{height: 100, width: 100}} />
+// 	}
+// })
+
+// var ProfileLink = React.createClass({
+// 	render: function () {
 // 		return (
-// 			<div> Hello {this.props.name} </div>
-// 		)
+// 			<div>
+// 			  <a href={'https://github.com/' + this.props.username}>
+// 			    {this.props.username}
+//               </a>
+//             </div>
+
+// 			)
 // 	}
 // });
 
-var ProfilePic = React.createClass({
-	render: function () {
-		return <img src={this.props.imageUrl} style={{height: 100, width: 100}} />
-	}
-})
+// var ProfileName = React.createClass({
+// 	render: function () {
+// 		return (
+// 			<div>{this.props.name}</div>
+// 	    )
+// 	}
+// });
 
-var ProfileLink = React.createClass({
-	render: function () {
-		return (
-			<div>
-			  <a href={'https://github.com/' + this.props.username}>
-			    {this.props.username}
-              </a>
-            </div>
+// var Avatar = React.createClass({
+// 	render: function () {
+// 		return (
+//           <div>
+//             <ProfilePic imageUrl={this.props.user.image} />
+//             <ProfileName name={this.props.user.name} />
+//             <ProfileLink username={this.props.user.username} />
+//           </div>
 
-			)
-	}
-});
-
-var ProfileName = React.createClass({
-	render: function () {
-		return (
-			<div>{this.props.name})</div>
-	    )
-	}
-});
-
-var Avatar = React.createClass({
-	render: function () {
-		return (
-          <div>
-            <ProfilePic imageUrl={this.props.image} />
-            <ProfileName name={this.props.name} />
-            <ProfileLink username={this.props.username} />
-          </div>
-
-		)
-	}
-})
+// 		)
+// 	}
+// })
 
 ReactDOM.render(
-  <Avatar use={USER_DATA}/>,
+  <Avatar user={USER_DATA} />,
   document.getElementById('app')
 );
